requirements.txt
Flask
azure-ai-documentintelligence
azure-core